refactor: NumberGeneratorStrategy -> domain, RandomNumberGenerator -> domain 바깥으로 이동
