+++ src/main/java/racingCar/domains/NumberGeneratorStrategy.java
@@ -1,4 +1,4 @@
-package racingCar.interfaces;
+package racingCar.domains;
 
 public interface NumberGeneratorStrategy {
     public int generate();
